Panic support.
This code needs to be very resilient: it's no good having something go
wrong halfway through te flight. So, implement a panic function, which
jumps back up to the top of loop(), from where we can try to reset
things and start again.

gps_setup moves into loop(), because it potentially calls panic, so it
has to be under the PANIC_CATCH. debug_setup does not, because panic
uses the debug functions, so the debug functions mustn't panic.

The implicit loop in the Arduino library's main() is now replaced with
an explicit loop in our loop(). This is necessary because of the way
setjmp works: any call to longjmp (ie. panic()) must be made before
the function which called setjmp (ie. PANIC_CATCH) returns. Otherwise
you get segfaults (or worse, probably, since we're running directly on
the hardware without a kernel).