A basic scheduler.
This system is going to have various tasks which need servicing at
different intervals. Create a task data structure to keep track of them,
which holds the time at which this task next needs attention. This makes
the main loop entirely generic: it just loops through the tasks,
handling the ones which need handling.