Less get these yur strengs into flash...
It seems that having more than one or two static buffers hanging around
is a very bad idea. So, instead of snprintfing into various buffers as
we need to, create a bprintf which formats a string into a single fixed
buffer. Also create Fprintf, which expects its format argument to be a
flash string, and strFdup, which copies a flash string into bprintf's
buffer.

Divide warn into three functions: warn takes a single flash string, and
prints it with a newline; warnf takes a flash format string and
arguments, and is rather less efficient; warnx takes a plain string, and
doesn't append a newline.