OK, that is *quite* enough of that.
Dear g++,

First you give me this:

    error: invalid conversion from ‘char*’ to ‘uint8_t* {aka unsigned
    char*}’ [-fpermissive]
    error:   initializing argument 2 of ‘uint8_t twi_readFrom(uint8_t,
    uint8_t*, uint8_t, uint8_t)’ [-fpermissive]

It's an ARRAY OF BYTES. Just PASS THE DAMN POINTER.

But OK. It is the wrong type, and all you actually want is a typecast
(and never mind that a type system that can be cast away like that isn't
worth a damn), so I can do that. But then you come up with:

    sorry, unimplemented: non-trivial designated initializers not
    supported

which makes NO SENSE to someone who isn't a C++ compiler writer, and
still makes ABSOLUTELY NO SENSE when it transpires that what you're
actually complaining about is this:

    ubx_cfg_prt foo = { .port = 0 };

because apparently you can't cope with structure initialisers unless the
elements are initialised in order from the beginning, and if I leave one
out you get all lost and confused. I know you're trying to do deeply
complicated things like find and run constructors and overloaded cast
operators and God knows what else, but I really wish you'd just STOP IT.
It's a STRUCTURE, it DOESN'T HAVE ANY OF THAT CRAP; fill in the bits I
gave you, zero the rest, and STOP COMPLAINING.

[I am converting as much as possible from C++ to C, because C++ is just
not suitable for programming at this level. As a bonus, just removing
Wire seems to have reduced the code size from 8k to 4k...]