Unify the flash and ram address spaces.
We've got less than 64k of addressable memory altogether, so the whole
lot can be addressed using a single 16-bit pointer. Set the top bit of a
pointer if an address is in flash, otherwise leave it clear.

These 'pointers' cannot be dereferenced in C++, because that means using
the __flash address space to generate the LPM instructions, and for some
reason that isn't supported in C++. But they can be generated in C++
(that's just arithmetic), and passed through, and so on, with no
problems.

It also turns out <Arduino.h> can be included from C (just not the other
headers), so I don't need cxxcompat.h any more.