Tidy up ubx.c after all that debugging.
I've gone back to my original version of pad_dump: I don't think there's
anything wrong with it. I don't know what was causing the dumps to
display incorrectly, but I have noticed that the ubx_pad structure I
added to ubx.c was too short (it only had one byte of payload), so maybe
that was the cause.

For the moment, replace upad with a static byte array. I don't like
this: the whole point of the pad was to avoid lots of static arrays. I
need to do two things here: change the logging functions so they write a
character at a time, and don't need a buffer; and implement some limited
form of dynamic allocation. (I don't want to just use malloc, even
though the C++ stuff is probably using it, because I suspect the
overhead and fragmentation will kill us once we need to allocate a 512b
buffer for writing to SD.)