twi will only send or recv 32 bytes at a time.
This means we have to loop over the buffer until we've done it all.

Conveniently, however, it seems the UBX module ignores I2C NACKs in the
middle of a packet: I had been assuming it would drop the rest of the
packet, or simply fail, if it got a NACK before it had finished sending.
This means that twi's other irritating behaviour of always sending a
NACK at the end of every receiveFrom is less important than it seemed,
and I can read a packet incrementally after all. This also means I can
drop packets when necessary.

Drop the suffix from ubx_{send,recv}_packet. Change the API for _recv:
the pkt->len member is now a maximum length, and the ->type field will
be checked if set. Drop ubx_send_with_reply, since it doesn't really do
anything useful.