amd64: CLOCK_VIRTUAL doesn't count the right time.
I was expecting that, since we do a lot of busy-waiting (which is Not
Good on a real system, but never mind), that CLOCK_VIRTUAL would
approximate realtime-passed-since-process-start. In fact, it doesn't
seem to increment at all, so just use REALTIME_FAST instead and subtract
the time of the first call.