Changes to build on Debian.
The fact '-V.PARSEDIR:?ok:' works on FreeBSD appears to be an FBSD
extension to bmake: FBSD's make traditionally fully-expanded variables
requested with -V, and this behaviour was kept for the FBSD port of
bmake, because ports uses -V all over the place. Other bmakes and
pmakes don't do this by default, and might or might not have -dV and
.MAKE.EXPAND_VARIABLES to control it.

Linux seems to keep clock_gettime in -lrt. Grr, I'm going to actually
have to do some proper configuring...