There's no need for 'make obj'.
I'm already relying on .OBJDIR, so I can just create the obj directory
immediately. I wasn't doing that before because I didn't have the full
list of OBJDIRS yet, but the cases are not the same: .OBJDIR needs to
exist now, so we can chdir into it, but the OBJDIRS only need to exist
before we run any targets, so .BEGIN will do fine for them.