local isn't in POSIX shell.
It seems to be an ash (BSD) extension. It is, sort-of, possible to
emulate local in strict POSIX shell; it's slow and it's ugly but it
seems to work, at least in ksh93 which is the only remotely-POSIX shell
I can find which doesn't support local already. ksh93 is also /bin/sh on
Illumos, and (AIUI) /usr/xpg4/bin/sh on Solaris versions that still have
a non-POSIX SVR4 shell as /bin/sh, so this would probably be worth
having in a properly portable configure script; but I'm not yet
convinced any of this is worth the effort for this project.

Annoyingly, ksh93 has its own local variables, using 'typeset', but they
only work in 'function foo { }' functions, not POSIX-style 'foo () { }'
functions.

ksh93 also gets very upset if you try to shift too far. Apparently this
is allowed.