Put in a delay when we get 0xff looking for sync.
It's not yet clear to me under what circumstances this UBX device will
return spurious 0xffs, but if we get an 0xff while looking for sync we
should definitely wait a bit rather than spamming the TWI bus with
requests. (Apart from anything else, driving the bus takes power.)

Ignore (and log) any unexpected bytes received while waiting for sync.
This doesn't mean we can just start looking for sync in the middle of a
packet, though, since the UBX protocol is binary and b5 62 is, for
example, 25269 as a LE 16-bit int. If we drop a packet (for instance, if
it's too long) we need to read to the end properly, throwing the bytes
away.

Since Wire has an unnecessary buffer, we might as well make use of it.
requestFrom will drop any unread data in the buffer, so we mustn't call
it when there is data available, but when we do call it we might as well get
as much as we can.