Make a start on the SD card code.
Unfortunately if we build in both the SD card stuff and the other tasks
we run out of memory, and the Arduino crashes. For now, comment out the
other tasks in this branch, so we can get the SD stuff working on its
own.

Later on we will have to start moving constant structures into flash,
which will be annoying from C++...

I've no idea if this will build from the IDE: I suspect it won't. At
least, it won't build a working executable, because it always compiles
everything in; it may not build at all, depending on where exactly it
looks for source files when its choosing what to compile.