Allow tasks to wait for multiple SWIs.
I didn't do this originally because I was worried that it introduces a
hard 16-SWI limit (because of the format of a wchan), but I think that's
less of a concern than the need to wait for multiple SWIs. I expect, for
instance, that the SD logger will want to wait for all the SWIs.